WINDSONIC/requirements.txt
pyserial==3.4
APScheduler==3.6.3
